chore(knockdog): API 관련 파일 shaerd/api로 모두 이동
