--- apps/knockdog/src/shared/lib/error/api-error.ts
@@ -1,11 +1,0 @@
-import { ApiErrorCode } from './error-code';
-
-export class ApiError extends Error {
-  constructor(
-    public code: ApiErrorCode,
-    public message: string
-  ) {
-    super(message);
-    this.name = 'ApiError';
-  }
-}
