--- apps/knockdog/src/shared/lib/error/index.ts
@@ -1,2 +1,0 @@
-export * from './api-error';
-export * from './error-code';
