--- apps/knockdog/src/shared/types/index.ts
@@ -1,1 +1,0 @@
-export type { ApiResponse } from './api';
